requirements.txt
Flask
requests
groq
matplotlib
pandas
beautifulsoup4
lxml
python-dotenv
streamlit